Add support for back link navigation (#16)
Add back link navigation. The process of converting the specification/graphql schema now follows a pipeline pattern. The GraphQLSchema is gradually transformed by tree visitors to add types, connection and reverse links. Types and fields are annotated with directives that the next visitors in the pipeline can recognize and act upon.